Add DNSBL support from charybdis.
DNSBLs are checked when the USER message is received and prevent completing
registration until done (this means that a CAP negotiation can be done in
parallel with DNSBL lookups). A pending DNSBL lookup also prevents the
unknown connection timing out. The result is checked at the same time as
klines. Reject cache applies to blacklists as to klines. 

There are no server notices, other than for invalid replies (not A or
outside 127.0.0.0/8).

The kline_exempt auth flag exempts from DNSBL checks, as does the new
dnsbl_exempt flag. In either case, a notice is still sent to warn the user.

The new /stats n shows blacklists and the number of hits since the last
rehash.

Since I do not know of any DNSBLs for IPv6, DNSBL checking is only done for
IPv4 clients.

Compared to charybdis, the preClient struct (to store things until
registration) is not present in ircd-ratbox, so I put the DNSBL fields in
localClient instead.

I have not ported variable substitution in the reason; the reason field is
just an exact string. Substitution can be useful to provide clickable URLs.


git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@26700 b93f080a-4cfa-0310-a19a-80a68f531ef9