Enable -Wformat checking on numerics.
The replies array is gone, all the texts are now #defines in a new file
include/messages.h. The form_str() function is now a macro that returns
one of the strings. It only accepts constants (usually a RPL_* or ERR_*
define from numeric.h).

The three places that passed a variable to form_str() are a bit uglier
now.

Various problems are now visible in -Wformat; this commit does not fix
them.


git-svn-id: http://svn.ratbox.org/svnroot/ircd-ratbox/trunk@27317 b93f080a-4cfa-0310-a19a-80a68f531ef9